# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Labor Connect API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ApiError {
  error: string;
}

export interface Laborer {
  id: number;
  name: string;
  phone: string;
  location: string;
  workType: string;
  /** @nullable */
  bio?: string | null;
  /** @nullable */
  dailyRate?: number | null;
  /** @nullable */
  lat?: number | null;
  /** @nullable */
  lng?: number | null;
  isAvailable: boolean;
  /** @nullable */
  distanceKm?: number | null;
  createdAt: string;
}

export interface LaborerInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  phone: string;
  /** @minLength 1 */
  location: string;
  /** @minLength 1 */
  workType: string;
  bio?: string;
  dailyRate?: number;
  lat?: number;
  lng?: number;
}

export interface LaborerUpdate {
  name?: string;
  phone?: string;
  location?: string;
  workType?: string;
  bio?: string;
  dailyRate?: number;
  lat?: number;
  lng?: number;
  isAvailable?: boolean;
}

export interface Conversation {
  id: number;
  laborerId: number;
  laborerName: string;
  hirerPhone: string;
  /** @nullable */
  lastMessage: string | null;
  updatedAt: string;
}

export interface ConversationInput {
  laborerId: number;
  hirerPhone: string;
}

export interface Message {
  id: number;
  conversationId: number;
  senderPhone: string;
  content: string;
  createdAt: string;
}

export interface MessageInput {
  senderPhone: string;
  content: string;
}

export type StatsTopWorkTypesItem = {
  workType: string;
  count: number;
};

export interface Stats {
  totalLaborers: number;
  availableLaborers: number;
  totalWorkTypes: number;
  topWorkTypes: StatsTopWorkTypesItem[];
}

export type ListLaborersParams = {
/**
 * Search by name, location, or work type
 */
search?: string;
workType?: string;
location?: string;
lat?: number;
lng?: number;
radiusKm?: number;
};

export type DeleteLaborerParams = {
/**
 * Must match the registered phone number
 */
phone: string;
};

export type ListConversationsParams = {
phone: string;
};